Add automated testing tags to linter
Hey, this is meant to make the linter aware of tags that are supported by the automated tester. The only tag used, at least for now, is `>>tester`, which takes an argument: `always`, `never`, `avoid`, `prefer`. Let me know if any of the changes in this commit are insufficient/wrong.